Further changes after D7237 to get OD compiling
git-svn-id: https://svn.opendtect.org/od/trunk@38397 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/EarthModel/emfault3dpainter.cc
@@ -277,7 +277,8 @@ bool Fault3DPainter::paintStickOnRLine( const Geometry::FaultStickSurface& fss,
 	bid = SI().transform( pos.coord() );
 	const TrcKey trckey = Survey::GM().traceKey(
 		Survey::GM().default3DSurvID(), bid.inl(), bid.crl() );
-	const int idx = path_->indexOf( trckey );
+	const int idx = (int)path_->indexOf( trckey );
+	    //TODO the cast above violates the design
 	if ( idx < 0 ) continue;
 
 	Coord3 editnormal( getNormalInRandLine(idx), 0 );
@@ -452,7 +453,8 @@ void Fault3DPainter::genIntersectionAuxData( EM::Fault3D& f3d,
 	{
 	    const TrcKey trckey = Survey::GM().traceKey(
 		    Survey::GM().default3DSurvID(),posbid.inl(),posbid.crl() );
-	    const int trcidx = path_->indexOf( trckey );
+	    const int trcidx = (int)path_->indexOf( trckey );
+		//TODO the cast above violates the design
 	    if ( trcidx != -1 )
 	    {
 		const double z = zat ? zat->transform(pos) : pos.z;
