+++ src/EarthModel/emfaultstickpainter.cc
@@ -150,7 +150,8 @@ bool FaultStickPainter::addPolyLine()
 			const BinID bid = SI().transform( pos.coord() );
 			const TrcKey trckey = Survey::GM().traceKey(
 			   Survey::GM().default3DSurvID(),bid.inl(),bid.crl() );
-			const int idx = path_->indexOf( trckey );
+			const int idx = (int)path_->indexOf( trckey );
+			    //TODO the cast above violates the design
 			if ( idx < 0 ) continue;
 
 			Coord3 editnormal( getNormalInRandLine(idx), 0 );
